+++ lib/positioning/advisory_lock.rb
@@ -1,35 +1,54 @@
+require "fileutils"
+
 module Positioning
   class AdvisoryLock
     Adapter = Struct.new(:aquire, :release, keyword_init: true)
 
-    def initialize(positioned)
-      @positioned = positioned
+    attr_reader :base_class
+
+    def initialize(base_class)
+      @base_class = base_class
 
       @adapters = {
         "Mysql2" => Adapter.new(
-          aquire: -> { "SELECT GET_LOCK(#{connection.quote(lock_name)}, -1)" },
-          release: -> { "SELECT RELEASE_LOCK(#{connection.quote(lock_name)})" }
+          aquire: -> { connection.execute "SELECT GET_LOCK(#{connection.quote(lock_name)}, -1)" },
+          release: -> { connection.execute "SELECT RELEASE_LOCK(#{connection.quote(lock_name)})" }
         ),
         "PostgreSQL" => Adapter.new(
-          aquire: -> { "SELECT pg_advisory_lock(#{lock_name.hex & 0x7FFFFFFFFFFFFFFF})" },
-          release: -> { "SELECT pg_advisory_unlock(#{lock_name.hex & 0x7FFFFFFFFFFFFFFF})" }
+          aquire: -> { connection.execute "SELECT pg_advisory_lock(#{lock_name.hex & 0x7FFFFFFFFFFFFFFF})" },
+          release: -> { connection.execute "SELECT pg_advisory_unlock(#{lock_name.hex & 0x7FFFFFFFFFFFFFFF})" }
+        ),
+        "SQLite" => Adapter.new(
+          aquire: -> {
+            FileUtils.mkdir_p "#{Dir.pwd}/tmp"
+            filename = "#{Dir.pwd}/tmp/#{lock_name}.lock"
+            FileUtils.touch filename
+            @file ||= File.open filename, "r+"
+            @file.flock File::LOCK_EX
+          },
+          release: -> {
+            @file.flock File::LOCK_UN
+          }
         )
       }
+
+      @adapters.default = Adapter.new(aquire: -> {}, release: -> {})
     end
 
-    def aquire
-      execute adapter.aquire.call if adapter
+    def aquire(record)
+      adapter.aquire.call
     end
 
-    def release
-      execute adapter.release.call if adapter
+    def release(record)
+      adapter.release.call
     end
+
+    alias_method :before_save, :aquire
+    alias_method :before_destroy, :aquire
+    alias_method :after_commit, :release
+    alias_method :after_rollback, :release
 
     private
-
-    def base_class
-      @positioned.class.base_class
-    end
 
     def connection
       base_class.connection
@@ -39,16 +58,16 @@ module Positioning
       connection.adapter_name
     end
 
-    def lock_name
-      ActiveSupport::Digest.hexdigest "#{connection.current_database}.#{base_class.table_name}"
-    end
-
     def adapter
       @adapters[adapter_name]
     end
 
-    def execute(command)
-      ActiveRecord::Base.connection.execute command
+    def lock_name
+      lock_name = ["positioning"]
+      lock_name << connection.current_database if connection.respond_to?(:current_database)
+      lock_name << base_class.table_name
+
+      ActiveSupport::Digest.hexdigest lock_name.join(".")
     end
   end
 end